// frontend/src/app/core/shared/common/enum/inventory-grn.enum.ts
export enum InventoryGrnEnum {
  grn_no = 'grn_no',
  grn_date = 'grn_date',
  source = 'source',
  purchase_order = 'purchase_order',
  vendor_bill = 'vendor_bill',
  vendor_bill_date = 'vendor_bill_date',
  vendor = 'vendor',
  vendor_name = 'vendor_name',
  vendor_address = 'vendor_address',
  note = 'note',
  batch_no = 'batch_no',
  other_ref_no = 'other_ref_no',
  vehicle_no = 'vehicle_no',
  arrive_date = 'arrive_date',
  exit_date = 'exit_date',
  remarks = 'remarks',
  grn_items = 'grn_items',
}

export enum GrnItemEnum {
  asset = 'asset',
  hs_code = 'hs_code',
  uom = 'uom',
  po_qty = 'po_qty',
  received_qty = 'received_qty',
  unit_price = 'unit_price',
  discount = 'discount',
  discount_amount = 'discount_amount',
  tax = 'tax',
  tax_amount = 'tax_amount',
  reject_qty = 'reject_qty',
  accepted_qty = 'accepted_qty',
  balance_qty = 'balance_qty',
  sub_total = 'sub_total',
  total_amount = 'total_amount',
  trc_number = 'trc_number',
  remarks = 'remarks',
}